ml/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
monai
nibabel
numpy
pydicom
simpleitk
